Cleaned up the solution and finished a work with captcha service

+++ ArgusBot.BL/Services/Implementation/CaptchaService.cs
@@ -1,5 +1,7 @@
 ﻿using ArgusBot.BL.Services.Interfaces;
 using ArgusBot.DAL.Models;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -13,11 +15,14 @@ namespace ArgusBot.BL.Services.Implementation
     {
         private readonly ITelegramBotClient _client;
         private readonly ICheckListService _checkListService;
-
-        public CaptchaService(ITelegramBotClient client, ICheckListService checkListService)
+        private readonly ILogger<ICaptchaService> _logger;
+        private readonly IConfiguration _config;
+        public CaptchaService(ITelegramBotClient client, ICheckListService checkListService, ILogger<ICaptchaService> logger, IConfiguration config)
         {
             _client = client;
             _checkListService = checkListService;
+            _logger = logger;
+            _config = config;
         }
         public async Task InitiateCaptchaProcess(IEnumerable<Telegram.Bot.Types.User> newUsers, long groupId)
         {
@@ -28,31 +33,34 @@ namespace ArgusBot.BL.Services.Implementation
                 {
                     int firstTerm = GetRandomTerm();
                     int secondTerm = GetRandomTerm();
+                    _logger.LogInformation($"Create a new captcha for user : {user.Id} in group {groupId}");
                     await _checkListService.CreateCheckAsync(groupId, user.Id, (firstTerm + secondTerm).ToString());
-                    var sentMessage = await _client.SendTextMessageAsync(groupId, 
+                    var sentMessage = await _client.SendTextMessageAsync(groupId,
                                                     $"{user.Username}, please write an answer for the following question: {firstTerm} + {secondTerm} = ?");
                     await _checkListService.UpdateQuestionMsgId(user.Id, sentMessage.MessageId);
                 }
             }
         }
-
         public async Task ProcessMesagge(Message message)
         {
-            IEnumerable<Check> checkNotes =  await _checkListService.GetAllFromCheckListAsync();
-            Check currentCheck = checkNotes.SingleOrDefault(u => u.UserId == message.From.Id);
+            Check currentCheck = await _checkListService.GetCheckForUser(message.From.Id);
             if (currentCheck != null)
             {
+                if (currentCheck.Status == StatusTypes.Successful)
+                {
+                    return;
+                }
                 var correctAnswer = currentCheck.CorrectAnswer;
                 if (message.Text == correctAnswer)
                 {
+                    _logger.LogInformation($"User {message.From.Id} has succesfully passed captcha!");
                     await _checkListService.ChangeCheckStatusForUserAsync(message.From.Id, (byte)StatusTypes.Successful);
                     await _client.DeleteMessageAsync(message.Chat.Id, currentCheck.QuestionMessageId);
                 }
                 await _client.DeleteMessageAsync(message.Chat.Id, message.MessageId);
+                return;
             }
-            return;
         }
-
         private int GetRandomTerm()
         {
             var rand = new Random();
